+++ CHANGELOG.md
@@ -5,6 +5,8 @@ UNRELEASED
   https://gitlab.com/gitlab-org/gitaly/merge_requests/332
 - Don't use 'bundle show' to find Linguist
   https://gitlab.com/gitlab-org/gitaly/merge_requests/339
+- Fix diff parsing when the last 10 bytes of a stream contain newlines
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/348
 
 v0.39.0
 - Reimplement FindAllTags RPC in Ruby